feat: much better implementation of custom cursor (widget)
